requirements.txt
pyrogram==1.4.16
tgcrypto<=1.2.3
aiohttp<=3.8.1
python-dotenv<=0.20.0
motor
aiofiles
dnspython
apscheduler
requests
pyshorteners
